Switch between normal loop and fast loop depending on IEC61851 state.
When we are in state C, the standard says that we have to turn the
contactor off within at most 100ms!
With the normal loop we have an ADC integration time of 250ms, so we
can't possibly react fast enough.
To fix this we have a "fast loop" that is used in state C.
The fast loop measures with an integration time of TBDms and it only
measures the voltage between CP/PE.
The voltage between PP/PE is ignored (the cable obviously can't be
changed out while a car is charging, so it is save to ignore the
PP/PE voltage while in state C.