Add special handling for ID.3 behavior
When an ID.3 is connected to the WARP charger and the duty cycle is already
below 100% (the wallbox is ready) but the contactor is not yet activated, the
ID.3 somtimes generates a spike in the resistance that we measure when it
engages the resistor to apply 880 ohm between CP/PE. We have not seen this in
other cars, we assume this is some kind of capacitive effect. To make sure
that we don't cancel the charging here, we increase the "STATE A" threshold and
the "infinite resistance" threshold for this scenario.